src/c1-data-scraping/requirements.txt
regex==2023.3.23
requests==2.28.1
beautifulsoup4==4.11.1
google-cloud-core==2.3.3
google-cloud-storage==2.11.0